requirements.txt
fastapi
uvicorn
aiogram==3.7.0
sqlalchemy[asyncio]
asyncpg
aiohttp
apscheduler
pydantic
python-dotenv
bcrypt